test(warp/hyperelastic/hess_quad): Refine numeric differentiation precision
Adjusts the epsilon value used for numerical differentiation in Hessian quadratic form tests from `1e-4` to `1e-5`. This change aims to improve the accuracy of the numerical approximation. The absolute tolerance for `assert_allclose` is also updated from `1e-10` to `1e-9` to allow for slight floating-point variations that may arise despite the increased precision of the numerical differentiation.